migrate CharSequenceToContainAtLeastExpectationsTest to kotlin-test
